Delete Item image file when appropriate
Since multer needs to come before the express-validator form validation
in the middleware chain, the image file was being uploaded on the Create
and Update routes even when the user input was invalid. To work around
this, I have added checks to see whether any errors occurred in the
validation for the Create and Update POST controllers and if so, the
uploaded file is deleted immediately. Also, updating an Item record from
having an image to not having an image wasn't deleting the existing
file, so I have made sure to delete the file in that case by checking
if the original Item record had an image before overwriting the record.